add tools chat_administrators
